Extract DismissDetectingModel
